% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcleanLite-package.R
\docType{package}
\name{textcleanLite}
\alias{package-textcleanLite}
\alias{textcleanLite}
\alias{textcleanLite-package}
\title{Text Cleaning Tools}
\description{
Tools to clean and process text.
}

